reorganize rpc primitives
