+++ primitives/rpc/src/lib.rs
@@ -1,7 +1,13 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
 pub mod api;
-pub mod types;
+pub mod chain;
+pub mod identity;
+pub mod transactions;
+pub mod verifier;
 
 pub use api::*;
-pub use types::*;
+pub use chain::*;
+pub use identity::*;
+pub use transactions::*;
+pub use verifier::*;
