--- primitives/rpc/src/types.rs
@@ -1,220 +1,0 @@
-use codec::{Decode, Encode};
-use scale_info::{
-	prelude::{string::String, vec::Vec},
-	TypeInfo,
-};
-#[cfg(feature = "std")]
-use serde::{Deserialize, Serialize};
-
-pub type ByPassToken = String;
-pub type PhoneNumberHash = sp_core::H512;
-
-#[derive(Encode, Decode, TypeInfo)]
-#[cfg_attr(feature = "std", derive(Debug, Serialize, Deserialize))]
-pub struct UserInfo<AccountId> {
-	pub account_id: AccountId,
-	pub nonce: u64,
-	pub user_name: String,
-	pub phone_number_hash: PhoneNumberHash,
-	pub balance: u64,
-	pub trait_scores: Vec<TraitScore>,
-	// pre_keys
-	pub karma_score: u32,
-	pub community_membership: Vec<CommunityMembership>,
-}
-
-#[derive(Encode, Decode, TypeInfo, PartialEq, Eq, PartialOrd, Ord)]
-#[cfg_attr(feature = "std", derive(Debug, Serialize, Deserialize))]
-pub struct TraitScore {
-	pub trait_id: u32,
-	pub karma_score: u32,
-	pub community_id: u32,
-}
-
-#[derive(Encode, Decode, TypeInfo, PartialEq, Eq, PartialOrd, Ord)]
-#[cfg_attr(feature = "std", derive(Debug, Serialize, Deserialize))]
-pub struct CommunityMembership {
-	pub community_id: u32,
-	pub karma_score: u32,
-	pub is_admin: bool,
-}
-
-#[derive(Encode, Decode, TypeInfo)]
-#[cfg_attr(feature = "std", derive(Debug, Serialize, Deserialize))]
-pub struct SignedTransaction<AccountId, Signature> {
-	pub signer: Option<AccountId>,
-	pub transaction_body: Vec<u8>,
-	pub signature: Option<Signature>,
-}
-
-#[derive(Encode, Decode, TypeInfo)]
-#[cfg_attr(feature = "std", derive(Debug, Serialize, Deserialize))]
-pub struct SignedTransactionWithStatus<AccountId, Signature, TransactionEvent> {
-	pub signed_transaction: SignedTransaction<AccountId, Signature>,
-	pub status: TransactionStatus,
-	pub from: Option<UserInfo<AccountId>>,
-	pub to: Option<UserInfo<AccountId>>,
-	pub timestamp: u64,
-	pub events: Vec<TransactionEvent>,
-	pub block_number: u32,
-	pub transaction_index: u32,
-}
-
-#[derive(Encode, Decode, TypeInfo, Default)]
-#[cfg_attr(feature = "std", derive(Debug, Serialize, Deserialize))]
-pub enum TransactionStatus {
-	#[default]
-	Unknown = 0_isize,
-	NotSubmitted,
-	Submitted,
-	Rejected,
-	OnChain,
-}
-
-#[derive(Encode, Decode, TypeInfo)]
-#[cfg_attr(feature = "std", derive(Debug, Serialize, Deserialize))]
-pub struct BlockchainStats {
-	/// Last block production time
-	pub last_block_time: u64,
-	/// Current block height
-	pub tip_height: u64,
-	/// Total number of executed transactions
-	pub transaction_count: u64,
-	/// Total number of payment transactions
-	pub payment_transaction_count: u64,
-	/// Total number of payment transactions with an appreciation
-	pub appreciations_transactions_count: u64,
-	/// Total number of payment transactions
-	pub update_user_transactions_count: u64,
-	/// Total number of verified user accounts
-	pub users_count: u64,
-	/// Total tx fees collected by block producers
-	pub fees_amount: u128,
-	/// Total number of kCents minted by the protocol since genesis
-	pub minted_amount: u128,
-	/// Total number of kCents in circulation by minting. Not including pre-mint
-	pub circulation: u128,
-	/// Total tx fee subsidies issued by the protocol
-	pub fee_subs_count: u64,
-	pub fee_subs_amount: u128,
-	pub signup_rewards_count: u64,
-	pub signup_rewards_amount: u128,
-	pub referral_rewards_count: u64,
-	pub referral_rewards_amount: u128,
-	pub validator_rewards_count: u64,
-	pub validator_rewards_amount: u128,
-	/// Amount of rewards paid to causes
-	pub causes_rewards_amount: u128,
-	// TODO: how can know KC cost?
-	// Estimated KC to USD exchange rate
-	// double exchange_rate = 19;
-}
-
-#[derive(Encode, Decode, TypeInfo)]
-#[cfg_attr(feature = "std", derive(Debug, Serialize, Deserialize))]
-pub struct GenesisData<AccountId> {
-	pub net_id: u32,
-	pub net_name: Vec<u8>,
-	pub genesis_time: u64,
-	pub signup_reward_phase1_alloc: u128,
-	pub signup_reward_phase2_alloc: u128,
-
-	pub signup_reward_phase1_amount: u128,
-	pub signup_reward_phase2_amount: u128,
-	pub signup_reward_phase3_start: u128,
-
-	pub referral_reward_phase1_alloc: u128,
-	pub referral_reward_phase2_alloc: u128,
-
-	pub referral_reward_phase1_amount: u128,
-	pub referral_reward_phase2_amount: u128,
-
-	pub tx_fee_subsidy_max_per_user: u64,
-	pub tx_fee_subsidies_alloc: u128,
-	pub tx_fee_subsidy_max_amount: u128,
-
-	pub block_reward_amount: u64,
-	pub block_reward_last_block: u64,
-
-	pub karma_reward_amount: u128,
-	pub karma_reward_alloc: u128,
-	pub karma_reward_top_n_users: u64,
-
-	// pub treasury_premint_amount: u64,
-	// pub treasury_account_id: AccountId,
-	// pub treasury_account_name: Vec<u8>,
-	pub char_traits: Vec<CharTrait>,
-	pub verifiers: Vec<PhoneVerifier<AccountId>>,
-}
-
-#[derive(Encode, Decode, TypeInfo)]
-#[cfg_attr(feature = "std", derive(Debug, Serialize, Deserialize))]
-pub struct CharTrait {
-	pub id: u32,
-	pub name: String,
-	pub emoji: String,
-}
-
-#[derive(Encode, Decode, TypeInfo)]
-#[cfg_attr(feature = "std", derive(Debug, Serialize, Deserialize))]
-pub struct PhoneVerifier<AccountId> {
-	pub account_id: AccountId,
-	pub name: String,
-}
-
-#[derive(Encode, Decode, TypeInfo)]
-#[cfg_attr(feature = "std", derive(Debug, Serialize, Deserialize))]
-pub struct Contact<AccountId> {
-	pub user_name: String,
-	pub account_id: AccountId,
-	pub phone_number_hash: PhoneNumberHash,
-	pub community_membership: Vec<CommunityMembership>,
-	pub trait_scores: Vec<TraitScore>,
-}
-
-#[derive(Encode, Decode, TypeInfo)]
-#[cfg_attr(feature = "std", derive(Debug, Serialize, Deserialize))]
-pub struct LeaderboardEntry<AccountId> {
-	pub user_name: String,
-	pub account_id: AccountId,
-	pub score: u32,
-	char_traits_ids: u32,
-}
-
-#[derive(Copy, Clone, Encode, Decode, TypeInfo)]
-#[cfg_attr(feature = "std", derive(Debug, Serialize, Deserialize))]
-pub enum VerificationResult {
-	Unspecified = 0,
-	/// There's already a user with the requested user name
-	UserNameTaken = 1,
-	/// User is verified using provided token
-	Verified = 2,
-	/// User is not verifier using provided token
-	Unverified = 3,
-	/// Request is missing required data
-	MissingData = 4,
-	/// Bad client signature
-	InvalidSignature = 5,
-	/// Different account associated with phone number
-	AccountMismatch = 6,
-}
-
-impl From<i32> for VerificationResult {
-	fn from(value: i32) -> Self {
-		match value {
-			0 => VerificationResult::Verified,
-			1 => VerificationResult::Unverified,
-			2 => VerificationResult::AccountMismatch,
-			_ => VerificationResult::Unspecified,
-		}
-	}
-}
-
-#[derive(Encode, Decode, TypeInfo)]
-#[cfg_attr(feature = "std", derive(Debug, Serialize, Deserialize))]
-pub struct VerificationEvidence<PublicKey, AccountId, Username, PhoneNumberHash> {
-	pub verifier_public_key: PublicKey,
-	pub account_id: AccountId,
-	pub username: Username,
-	pub phone_number_hash: PhoneNumberHash,
-}
